refactor:#265 client report 리팩토링
