+++ src/app/(client)/report/_components/bar-chart-skeleton.tsx
